fix(searcher-rs): decode tokens.decimals as i16 not i32 (SQL INT2/SMALLINT)
ROOT CAUSE of the system-wide data void. pool_sync_worker crashed at boot on
bootstrap_token_cache with:
  "error decoding column 2: mismatched types; Rust Option<i32> (INT4)
   not compatible with SQL INT2"

tokens.decimals is SMALLINT (INT2) in the schema; sqlx maps INT2→i16, INT4→i32.
Decoding as Option<i32> is a hard type mismatch → the worker terminated BEFORE
writing arbx:tokens / arbx:pool_reserves / arbx:v3_slot0. That single crash
cascaded into the entire zero-pipeline:
  - price_worker.discover_tokens joins allowlist against arbx:tokens → 0 tokens
    → attempted:0 → arbx:token_prices empty → no_price_oracle (87.8% of rejects)
  - reserves cache empty → reserves_cache_miss (11.1%)
  - build_market_state → None → the 31 math operators never fire
  - expected_profit_usd NULL → 0 viable opportunities

fee_tier in the sibling query IS integer (INT4) → Option<i32> stays correct there.
`decimals as u8` downstream is unchanged (i16 → u8 cast is identical).

Verified: cargo check --bin searcher-rs clean. pools_loaded:26 then crash → after
fix, worker should bootstrap caches fully.

+++ backend/searcher-rs/src/workers/pool_sync_worker.rs
@@ -929,7 +929,12 @@ impl PoolSyncWorker {
         // decimals must NOT crash the whole worker at boot (that bug left every reserves/
         // slot0 cache empty). Fail-honest: skip the unusable token + count it, never
         // fabricate a symbol/decimals for a token that can't be priced or matched.
-        let rows = sqlx::query_as::<_, (String, Option<String>, Option<i32>, Option<bool>)>(
+        // `decimals` is SMALLINT (INT2) in the tokens table — sqlx maps INT2 to i16,
+        // NOT i32. Decoding as Option<i32> crashes the whole worker at boot with
+        // "mismatched types; Rust Option<i32> (INT4) not compatible with SQL INT2",
+        // which left every reserves/slot0/token cache empty (cascade: no_price_oracle,
+        // reserves_cache_miss, math_evidence dead, 0 viable opportunities).
+        let rows = sqlx::query_as::<_, (String, Option<String>, Option<i16>, Option<bool>)>(
             r#"SELECT address, symbol, decimals, is_stablecoin
                FROM tokens WHERE chain_id = $1 AND is_active = TRUE"#,
         )
